+++ Circle_2/cub3d/srcs/draw_fov/etc/get_sky_image_data.c
@@ -6,11 +6,11 @@
 /*   By: hyunjuyo <hyunjuyo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/04 17:52:54 by hyunjuyo          #+#    #+#             */
-/*   Updated: 2021/04/08 14:14:03 by hyunjuyo         ###   ########.fr       */
+/*   Updated: 2021/04/08 16:34:32 by hyunjuyo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub3d.h"
+#include "cub3d_bonus.h"
 
 int		get_sky_color(t_game *game, int h, int w)
 {
@@ -38,7 +38,9 @@ void	get_sky_image_data(t_game *game)
 			"./textures/sky.xpm", &game->c_img.width, &game->c_img.height)))
 	{
 		printf("Error\n[sky]mlx_xpm_file_to_image() failed\n");
-		system("killall afplay");
+		waitpid(game->pid, &game->pid_status, WNOHANG);
+		if (game->pid_status == -1)
+			system("killall afplay");
 		exit(1);
 	}
 	game->c_img.data = (int *)mlx_get_data_addr(game->c_img.img,
