+++ Circle_2/cub3d/srcs/draw_fov/etc/texture_floor_bonus.c
@@ -6,7 +6,7 @@
 /*   By: hyunjuyo <hyunjuyo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/27 15:15:02 by hyunjuyo          #+#    #+#             */
-/*   Updated: 2021/04/05 17:31:59 by hyunjuyo         ###   ########.fr       */
+/*   Updated: 2021/04/08 16:18:31 by hyunjuyo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,7 +78,9 @@ void	texture_floor(t_game *game)
 	if (!game->f_img.img)
 	{
 		printf("Error\n[floor]mlx_xpm_file_to_image() failed\n");
-		system("killall afplay");
+		waitpid(game->pid, &game->pid_status, WNOHANG);
+		if (game->pid_status == -1)
+			system("killall afplay");
 		exit(1);
 	}
 	game->f_img.data = (int *)mlx_get_data_addr(game->f_img.img,
