+++ Circle_2/cub3d/srcs/keyset_n_utils/exit_game_bonus.c
@@ -1,21 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exit_game.c                                        :+:      :+:    :+:   */
+/*   exit_game_bonus.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hyunjuyo <hyunjuyo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/02 16:21:12 by hyunjuyo          #+#    #+#             */
-/*   Updated: 2021/04/04 15:56:33 by hyunjuyo         ###   ########.fr       */
+/*   Updated: 2021/04/08 16:32:56 by hyunjuyo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub3d.h"
+#include "cub3d_bonus.h"
 
 int		exit_game(t_game *game)
 {
 	mlx_destroy_window(game->mlx, game->win);
-	system("killall afplay");
+	waitpid(game->pid, &game->pid_status, WNOHANG);
+	if (game->pid_status == -1)
+		system("killall afplay");
 	exit(0);
 	return (0);
 }
